--- src/main/java/roomescape/TimeRepository.java
@@ -1,14 +1,0 @@
-package roomescape;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class TimeRepository {
-    private final JdbcTemplate jdbcTemplate;
-
-    public TimeRepository(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-}
